refactor governance fixtures
