+++ packages/scripts/tasks/fixture.js
@@ -1,99 +1,10 @@
 const { randomInt } = require('crypto');
 const { task } = require('hardhat/config');
-const inquirer = require('inquirer');
-const types = require('@synthetixio/core-js/utils/hardhat/argument-types');
-const { formatDate } = require('@synthetixio/core-js/utils/misc/dates');
 const logger = require('@synthetixio/core-js/utils/io/logger');
-const getPackageProxy = require('../internal/get-package-proxy');
-const getPeriodDate = require('../internal/get-period-date');
-const getTimestamp = require('../internal/get-timestamp');
+const types = require('@synthetixio/core-js/utils/hardhat/argument-types');
 const assertPeriod = require('../internal/assert-period');
-const { COUNCILS, ElectionPeriod } = require('../internal/constants');
-
-task('governance-fixtures', 'CLI tools for managing governance fixtures')
-  .addOptionalParam('instance', 'Deployment instance name', 'official', types.alphanumeric)
-  .setAction(async ({ instance }, hre) => {
-    //eslint-disable-next-line no-constant-condition
-    while (true) {
-      const councils = await initCouncils(hre, instance);
-
-      const ui = new inquirer.ui.BottomBar();
-
-      const timestamp = await getTimestamp(hre);
-      const blockNumber = await hre.ethers.provider.getBlockNumber();
-
-      const time = new Date(timestamp * 1000);
-      ui.log.write(
-        toOneLine({
-          Network: hre.network.name,
-          BlockNumber: blockNumber,
-          Timestamp: timestamp,
-          Date: formatDate(time),
-        })
-      );
-
-      const choices = [];
-
-      await asyncForEach(councils, async (council) => {
-        const councilChoices = [];
-
-        councilChoices.push(
-          new inquirer.Separator(
-            `${council.name.toLocaleUpperCase().replace('-', ' ')} (Period: ${
-              council.currentPeriod
-            })`
-          )
-        );
-
-        if (council.currentPeriod === 'Evaluation') {
-          councilChoices.push({
-            name: '  Evaluate election',
-            value: {
-              type: 'run',
-              name: 'evaluate-election',
-              args: { instance, council: council.name },
-            },
-          });
-        } else if (council.currentPeriod === 'Nomination') {
-          councilChoices.push({
-            name: '  Nominate 14 fixture candidates',
-            value: {
-              type: 'run',
-              name: 'fixture:candidates',
-              args: { instance, council: council.name, nominateAmount: '14' },
-            },
-          });
-        }
-
-        if (council.currentPeriod !== 'Evaluation') {
-          const nextPeriod = getNext(Object.keys(ElectionPeriod), council.currentPeriod);
-          const nextPeriodDate = await getPeriodDate(council.Proxy, nextPeriod);
-          councilChoices.push({
-            name: `  Forward to next period "${nextPeriod}" (timestamp: ${nextPeriodDate})`,
-            value: {
-              type: 'run',
-              name: 'fast-forward-to',
-              args: { instance, council: council.name, period: nextPeriod },
-            },
-          });
-        }
-
-        choices.push(...councilChoices);
-      });
-
-      const { response } = await inquirer.prompt({
-        type: 'list',
-        name: 'response',
-        message: 'What action do you want to perform?',
-        choices,
-      });
-
-      if (response.type === 'run') {
-        await hre.run(response.name, response.args);
-        continue;
-      }
-    }
-  });
+const getPackageProxy = require('../internal/get-package-proxy');
+const { COUNCILS } = require('../internal/constants');
 
 task('fixture:wallets', 'Create fixture wallets')
   .addOptionalParam('amount', 'Amount of wallets to fixture', '50', types.int)
@@ -111,12 +22,12 @@ task('fixture:wallets', 'Create fixture wallets')
         params: [wallets.map((w) => w.address), '0x10000000000000000000000'],
       });
     } else if (['local', 'localhost', 'hardhat'].includes(hre.network.name)) {
-      await asyncForEach(wallets, ({ address }) =>
-        hre.network.provider.request({
+      for (const { address } of wallets) {
+        await hre.network.provider.request({
           method: 'hardhat_setBalance',
           params: [address, '0x10000000000000000000000'],
-        })
-      );
+        });
+      }
     }
 
     for (const [i, wallet] of wallets.entries()) {
@@ -130,12 +41,7 @@ task('fixture:wallets', 'Create fixture wallets')
 
 task('fixture:candidates', 'Create fixture candidate nominations')
   .addOptionalParam('instance', 'Deployment instance name', 'official', types.alphanumeric)
-  .addParam(
-    'council',
-    'From which council to take the period time',
-    undefined,
-    types.oneOf(...COUNCILS)
-  )
+  .addParam('council', 'Target council deployment', undefined, types.oneOf(...COUNCILS))
   .addOptionalParam('nominateAmount', 'Amount of candidates to fixture', '14', types.int)
   .addOptionalParam('withdrawAmount', 'Amount of candidates to withdraw nomination', '2', types.int)
   .setAction(async ({ instance, council, nominateAmount, withdrawAmount }, hre) => {
@@ -170,12 +76,7 @@ task('fixture:candidates', 'Create fixture candidate nominations')
 
 task('fixture:votes', 'Create fixture votes to nominated candidates')
   .addOptionalParam('instance', 'Deployment instance name', 'official', types.alphanumeric)
-  .addParam(
-    'council',
-    'From which council to take the period time',
-    undefined,
-    types.oneOf(...COUNCILS)
-  )
+  .addParam('council', 'Target council deployment', undefined, types.oneOf(...COUNCILS))
   .addOptionalParam('amount', 'Amount of voters to fixture', '20', types.int)
   .addOptionalParam('ballotSize', 'Amount of cadidates for each ballot', '1', types.int)
   .setAction(async ({ instance, council, amount, ballotSize }, hre) => {
@@ -232,30 +133,6 @@ task('fixture:votes', 'Create fixture votes to nominated candidates')
     }
   });
 
-async function initCouncils(hre, instance) {
-  return await Promise.all(
-    COUNCILS.map(async (name) => {
-      const council = { name };
-      council.Proxy = await getPackageProxy(hre, name, instance);
-      council.currentPeriod = Object.keys(ElectionPeriod)[await council.Proxy.getCurrentPeriod()];
-      return council;
-    })
-  );
-}
-
-function getNext(arr, curr) {
-  if (!Array.isArray(arr)) throw new Error('Expected array value');
-  const currIndex = arr.findIndex((val) => val === curr);
-  if (currIndex === -1) throw new Error(`Item "${curr}" not found in "${JSON.stringify(arr)}"`);
-  return arr[currIndex === arr.length - 1 ? 0 : currIndex + 1];
-}
-
-function toOneLine(obj = {}) {
-  return Object.entries(obj)
-    .map(([key, val]) => `${key}: ${val}`)
-    .join(' | ');
-}
-
 function createArray(length = 0) {
   return Array.from(Array(Number(length)));
 }
@@ -270,8 +147,4 @@ function pickRand(arr, amount = 1) {
   }
 
   return res;
-}
-
-async function asyncForEach(arr, fn) {
-  return await Promise.all(arr.map(fn));
 }
